# JIT: slightly more aggressive tail duplication