Generalize tail duplication to catch patterns like the one in #37904 where
a trinary compare feeds a binary compare.