@AndyAyersMS 2021-11-03
cc @dotnet/jit-contrib 

Not sure this is in its final form yet; thinking about ways to generalize the pattern match. Would perhaps be a nice test case for a high-level declarative syntax driven match. Also might want to experiment with making this even more aggressive.

On the example from #37904 this now gets us the same codegen for `Compare1` and `Compare2`. For example:

```asm
;; before

; Assembly listing for method C`1:Compare1(Span`1,int):int
; Emitting BLENDED_CODE for X64 CPU with AVX - Windows
; optimized code
; rsp based frame
; partially interruptible
; No PGO data
; 0 inlinees with PGO data; 0 single block inlinees; 2 inlinees without PGO data
; Final local variable assignments
;
;  V00 arg0         [V00,T00] (  4,  8   )   byref  ->  rcx         ld-addr-op single-def
;  V01 arg1         [V01,T01] (  4,  3.50)     int  ->  rdx         single-def
;  V02 OutArgs      [V02    ] (  1,  1   )  lclBlk (32) [rsp+00H]   "OutgoingArgSpace"
;* V03 tmp1         [V03,T04] (  0,  0   )    bool  ->  zero-ref    "Inline return value spill temp"
;  V04 tmp2         [V04,T02] (  3,  5   )     int  ->  rax         ld-addr-op "Inlining Arg"
;  V05 tmp3         [V05,T03] (  4,  2.50)     int  ->  rcx         "Inline return value spill temp"
;* V06 tmp4         [V06    ] (  0,  0   )   byref  ->  zero-ref    V08._pointer(offs=0x00) P-INDEP "field V00._pointer (fldOffset=0x0)"
;* V07 tmp5         [V07    ] (  0,  0   )     int  ->  zero-ref    V08._length(offs=0x08) P-INDEP "field V00._length (fldOffset=0x8)"
;* V08 tmp6         [V08    ] (  0,  0   )  struct (16) zero-ref    "Promoted implicit byref"
;
; Lcl frame size = 40

G_M38094_IG01:
       sub      rsp, 40
						;; bbWeight=1    PerfScore 0.25
G_M38094_IG02:
       cmp      dword ptr [rcx+8], 0
       jbe      SHORT G_M38094_IG11
       mov      rax, bword ptr [rcx]
       mov      eax, dword ptr [rax]
       cmp      eax, edx
       jge      SHORT G_M38094_IG04
						;; bbWeight=1    PerfScore 9.25
G_M38094_IG03:
       mov      ecx, -1
       jmp      SHORT G_M38094_IG06
						;; bbWeight=0.50 PerfScore 1.12
G_M38094_IG04:
       cmp      eax, edx
       jle      SHORT G_M38094_IG05
       mov      ecx, 1
       jmp      SHORT G_M38094_IG06
						;; bbWeight=0.50 PerfScore 1.75
G_M38094_IG05:
       xor      ecx, ecx
						;; bbWeight=0.50 PerfScore 0.12
G_M38094_IG06:
       test     ecx, ecx
       jl       SHORT G_M38094_IG09
						;; bbWeight=1    PerfScore 1.25
G_M38094_IG07:
       xor      eax, eax
						;; bbWeight=0.50 PerfScore 0.12
G_M38094_IG08:
       add      rsp, 40
       ret      
						;; bbWeight=0.50 PerfScore 0.62
G_M38094_IG09:
       mov      eax, 1
						;; bbWeight=0.50 PerfScore 0.12
G_M38094_IG10:
       add      rsp, 40
       ret      
						;; bbWeight=0.50 PerfScore 0.62
G_M38094_IG11:
       call     CORINFO_HELP_RNGCHKFAIL
       int3     
						;; bbWeight=0    PerfScore 0.00

;; after

; Assembly listing for method C`1:Compare1(Span`1,int):int
; Emitting BLENDED_CODE for X64 CPU with AVX - Windows
; optimized code
; rsp based frame
; partially interruptible
; No PGO data
; 0 inlinees with PGO data; 0 single block inlinees; 2 inlinees without PGO data
; Final local variable assignments
;
;  V00 arg0         [V00,T00] (  4,  8   )   byref  ->  rcx         ld-addr-op single-def
;  V01 arg1         [V01,T01] (  3,  3   )     int  ->  rdx         single-def
;  V02 OutArgs      [V02    ] (  1,  1   )  lclBlk (32) [rsp+00H]   "OutgoingArgSpace"
;* V03 tmp1         [V03,T03] (  0,  0   )    bool  ->  zero-ref    "Inline return value spill temp"
;  V04 tmp2         [V04,T02] (  2,  4   )     int  ->  rax         ld-addr-op "Inlining Arg"
;* V05 tmp3         [V05,T04] (  0,  0   )     int  ->  zero-ref    "Inline return value spill temp"
;* V06 tmp4         [V06    ] (  0,  0   )   byref  ->  zero-ref    V08._pointer(offs=0x00) P-INDEP "field V00._pointer (fldOffset=0x0)"
;* V07 tmp5         [V07    ] (  0,  0   )     int  ->  zero-ref    V08._length(offs=0x08) P-INDEP "field V00._length (fldOffset=0x8)"
;* V08 tmp6         [V08    ] (  0,  0   )  struct (16) zero-ref    "Promoted implicit byref"
;
; Lcl frame size = 40

G_M38094_IG01:
       sub      rsp, 40
						;; bbWeight=1    PerfScore 0.25
G_M38094_IG02:
       cmp      dword ptr [rcx+8], 0
       jbe      SHORT G_M38094_IG07
       mov      rax, bword ptr [rcx]
       mov      eax, dword ptr [rax]
       cmp      eax, edx
       jl       SHORT G_M38094_IG05
						;; bbWeight=1    PerfScore 9.25
G_M38094_IG03:
       xor      eax, eax
						;; bbWeight=0.50 PerfScore 0.12
G_M38094_IG04:
       add      rsp, 40
       ret      
						;; bbWeight=0.50 PerfScore 0.62
G_M38094_IG05:
       mov      eax, 1
						;; bbWeight=0.50 PerfScore 0.12
G_M38094_IG06:
       add      rsp, 40
       ret      
						;; bbWeight=0.50 PerfScore 0.62
G_M38094_IG07:
       call     CORINFO_HELP_RNGCHKFAIL
       int3     
						;; bbWeight=0    PerfScore 0.00
```
SPMI on (on top of #61023) shows modest number of hits..

## aspnet.run.windows.x64.checked.mch:

```

Summary of Code Size diffs:
(Lower is better)

Total bytes of base: 13670369 (overridden on cmd)
Total bytes of diff: 13670229 (overridden on cmd)
Total bytes of delta: -140 (-0.00 % of base)
    diff is an improvement.
    relative diff is an improvement.
```
<details>

<summary>Detail diffs</summary>

```


Top file improvements (bytes):
         -20 : 28896.dasm (-5.25% of base)
         -20 : 14131.dasm (-5.14% of base)
         -20 : 4959.dasm (-5.14% of base)
         -18 : 38900.dasm (-3.38% of base)
         -16 : 14124.dasm (-1.06% of base)
         -16 : 4947.dasm (-1.06% of base)
         -12 : 28894.dasm (-0.81% of base)
          -8 : 18570.dasm (-0.87% of base)
          -8 : 24623.dasm (-0.86% of base)
          -2 : 6134.dasm (-0.81% of base)

10 total files with Code Size differences (10 improved, 0 regressed), 0 unchanged.

Top method improvements (bytes):
         -20 (-5.25% of base) : 28896.dasm - SocketAsyncEventArgs:StartOperationAccept():this
         -20 (-5.14% of base) : 14131.dasm - SocketAsyncEventArgs:StartOperationAccept():this
         -20 (-5.14% of base) : 4959.dasm - SocketAsyncEventArgs:StartOperationAccept():this
         -18 (-3.38% of base) : 38900.dasm - ExpressionUtils:RequiresCanRead(Expression,String,int)
         -16 (-1.06% of base) : 14124.dasm - Socket:AcceptAsync(SocketAsyncEventArgs,CancellationToken):bool:this
         -16 (-1.06% of base) : 4947.dasm - Socket:AcceptAsync(SocketAsyncEventArgs,CancellationToken):bool:this
         -12 (-0.81% of base) : 28894.dasm - Socket:AcceptAsync(SocketAsyncEventArgs,CancellationToken):bool:this
          -8 (-0.87% of base) : 18570.dasm - RelationalPropertyExtensions:GetColumnName(IReadOnlyProperty,byref):String
          -8 (-0.86% of base) : 24623.dasm - RelationalPropertyExtensions:GetColumnName(IReadOnlyProperty,byref):String
          -2 (-0.81% of base) : 6134.dasm - SocketAsyncEventArgs:StartOperationAccept():this

Top method improvements (percentages):
         -20 (-5.25% of base) : 28896.dasm - SocketAsyncEventArgs:StartOperationAccept():this
         -20 (-5.14% of base) : 14131.dasm - SocketAsyncEventArgs:StartOperationAccept():this
         -20 (-5.14% of base) : 4959.dasm - SocketAsyncEventArgs:StartOperationAccept():this
         -18 (-3.38% of base) : 38900.dasm - ExpressionUtils:RequiresCanRead(Expression,String,int)
         -16 (-1.06% of base) : 14124.dasm - Socket:AcceptAsync(SocketAsyncEventArgs,CancellationToken):bool:this
         -16 (-1.06% of base) : 4947.dasm - Socket:AcceptAsync(SocketAsyncEventArgs,CancellationToken):bool:this
          -8 (-0.87% of base) : 18570.dasm - RelationalPropertyExtensions:GetColumnName(IReadOnlyProperty,byref):String
          -8 (-0.86% of base) : 24623.dasm - RelationalPropertyExtensions:GetColumnName(IReadOnlyProperty,byref):String
         -12 (-0.81% of base) : 28894.dasm - Socket:AcceptAsync(SocketAsyncEventArgs,CancellationToken):bool:this
          -2 (-0.81% of base) : 6134.dasm - SocketAsyncEventArgs:StartOperationAccept():this

10 total methods with Code Size differences (10 improved, 0 regressed), 0 unchanged.

```

</details>

--------------------------------------------------------------------------------

## benchmarks.run.windows.x64.checked.mch:

```

Summary of Code Size diffs:
(Lower is better)

Total bytes of base: 7151127 (overridden on cmd)
Total bytes of diff: 7150968 (overridden on cmd)
Total bytes of delta: -159 (-0.00 % of base)
    diff is an improvement.
    relative diff is an improvement.
```
<details>

<summary>Detail diffs</summary>

```


Top file regressions (bytes):
           2 : 11056.dasm (0.43% of base)

Top file improvements (bytes):
         -18 : 6098.dasm (-1.36% of base)
         -16 : 18739.dasm (-0.87% of base)
         -15 : 12882.dasm (-2.74% of base)
         -14 : 18084.dasm (-1.82% of base)
         -13 : 12880.dasm (-1.06% of base)
         -13 : 9473.dasm (-2.11% of base)
         -13 : 21133.dasm (-3.01% of base)
         -13 : 12412.dasm (-0.89% of base)
         -13 : 6816.dasm (-3.10% of base)
          -8 : 18760.dasm (-4.21% of base)
          -7 : 13506.dasm (-8.86% of base)
          -5 : 13383.dasm (-0.94% of base)
          -4 : 1617.dasm (-0.13% of base)
          -3 : 18343.dasm (-0.08% of base)
          -2 : 4913.dasm (-0.80% of base)
          -2 : 13451.dasm (-0.50% of base)
          -2 : 18655.dasm (-0.21% of base)

18 total files with Code Size differences (17 improved, 1 regressed), 0 unchanged.

Top method regressions (bytes):
           2 ( 0.43% of base) : 11056.dasm - StateMachineBox`1[Int32,<ReadAsyncSlowPath>d__39][System.Int32,System.IO.Strategies.BufferedFileStreamStrategy+<ReadAsyncSlowPath>d__39]:ExecutionContextCallback(System.Object)

Top method improvements (bytes):
         -18 (-1.36% of base) : 6098.dasm - JetStream.Statistics:findOptimalSegmentationInternal(System.Single[][],System.Int32[][],System.Double[],JetStream.SampleVarianceUpperTriangularMatrix,int)
         -16 (-0.87% of base) : 18739.dasm - Microsoft.CodeAnalysis.CSharp.DataFlowPass:Scan(byref):System.Collections.Immutable.ImmutableArray`1[[Microsoft.CodeAnalysis.CSharp.PreciseAbstractFlowPass`1+PendingBranch[[Microsoft.CodeAnalysis.CSharp.DataFlowPass+LocalState, Microsoft.CodeAnalysis.CSharp, Version=2.10.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35]], Microsoft.CodeAnalysis.CSharp, Version=2.10.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35]]:this
         -15 (-2.74% of base) : 12882.dasm - Newtonsoft.Json.Serialization.JsonSerializerInternalReader:CreateNewList(Newtonsoft.Json.JsonReader,Newtonsoft.Json.Serialization.JsonArrayContract,byref):System.Collections.IList:this
         -14 (-1.82% of base) : 18084.dasm - Microsoft.CodeAnalysis.CSharp.Symbols.Metadata.PE.PEParameterSymbol:.ctor(Microsoft.CodeAnalysis.CSharp.Symbols.Metadata.PE.PEModuleSymbol,Microsoft.CodeAnalysis.CSharp.Symbol,int,bool,Microsoft.CodeAnalysis.CSharp.Symbols.TypeSymbol,System.Reflection.Metadata.ParameterHandle,int,byref):this
         -13 (-3.10% of base) : 6816.dasm - Newtonsoft.Json.Serialization.DefaultContractResolver:InitializeContract(Newtonsoft.Json.Serialization.JsonContract):this
         -13 (-0.89% of base) : 12412.dasm - Newtonsoft.Json.Serialization.JsonDictionaryContract:.ctor(System.Type):this
         -13 (-1.06% of base) : 12880.dasm - Newtonsoft.Json.Serialization.JsonSerializerInternalReader:CreateList(Newtonsoft.Json.JsonReader,System.Type,Newtonsoft.Json.Serialization.JsonContract,Newtonsoft.Json.Serialization.JsonProperty,System.Object,System.String):System.Object:this
         -13 (-3.01% of base) : 21133.dasm - Newtonsoft.Json.Serialization.JsonSerializerInternalReader:CreateNewDictionary(Newtonsoft.Json.JsonReader,Newtonsoft.Json.Serialization.JsonDictionaryContract,byref):System.Collections.IDictionary:this
         -13 (-2.11% of base) : 9473.dasm - System.Net.Http.HttpWindowsProxy:TryCreate(byref):bool
          -8 (-4.21% of base) : 18760.dasm - Microsoft.CodeAnalysis.CSharp.DataFlowPass:IntersectWith(byref,byref):bool:this
          -7 (-8.86% of base) : 13506.dasm - System.Net.Security.SslStream:get_IsAuthenticated():bool:this
          -5 (-0.94% of base) : 13383.dasm - System.Net.Security.SslStream:ValidateCreateContext(System.Net.Security.SslClientAuthenticationOptions,System.Net.Security.RemoteCertificateValidationCallback,System.Net.Security.LocalCertSelectionCallback):this
          -4 (-0.13% of base) : 1617.dasm - ProtoBuf.Meta.MetaType:BuildSerializer():ProtoBuf.Internal.Serializers.IProtoTypeSerializer:this
          -3 (-0.08% of base) : 18343.dasm - Microsoft.CodeAnalysis.CSharp.Binder:BindNamespaceOrTypeOrAliasSymbol(Microsoft.CodeAnalysis.CSharp.Syntax.ExpressionSyntax,Microsoft.CodeAnalysis.DiagnosticBag,Roslyn.Utilities.ConsList`1[[Microsoft.CodeAnalysis.CSharp.Symbol, Microsoft.CodeAnalysis.CSharp, Version=2.10.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35]],bool):Microsoft.CodeAnalysis.CSharp.Symbol:this
          -2 (-0.21% of base) : 18655.dasm - Microsoft.CodeAnalysis.CSharp.ConversionsBase:ClassifyImplicitBuiltInConversionFromExpression(Microsoft.CodeAnalysis.CSharp.BoundExpression,Microsoft.CodeAnalysis.CSharp.Symbols.TypeSymbol,Microsoft.CodeAnalysis.CSharp.Symbols.TypeSymbol,byref):Microsoft.CodeAnalysis.CSharp.Conversion:this
          -2 (-0.50% of base) : 13451.dasm - System.Net.Security.SslStream:ValidateCreateContext(System.Net.Security.SslAuthenticationOptions):this
          -2 (-0.80% of base) : 4913.dasm - System.Net.Sockets.SocketAsyncEventArgs:StartOperationAccept():this

Top method regressions (percentages):
           2 ( 0.43% of base) : 11056.dasm - StateMachineBox`1[Int32,<ReadAsyncSlowPath>d__39][System.Int32,System.IO.Strategies.BufferedFileStreamStrategy+<ReadAsyncSlowPath>d__39]:ExecutionContextCallback(System.Object)

Top method improvements (percentages):
          -7 (-8.86% of base) : 13506.dasm - System.Net.Security.SslStream:get_IsAuthenticated():bool:this
          -8 (-4.21% of base) : 18760.dasm - Microsoft.CodeAnalysis.CSharp.DataFlowPass:IntersectWith(byref,byref):bool:this
         -13 (-3.10% of base) : 6816.dasm - Newtonsoft.Json.Serialization.DefaultContractResolver:InitializeContract(Newtonsoft.Json.Serialization.JsonContract):this
         -13 (-3.01% of base) : 21133.dasm - Newtonsoft.Json.Serialization.JsonSerializerInternalReader:CreateNewDictionary(Newtonsoft.Json.JsonReader,Newtonsoft.Json.Serialization.JsonDictionaryContract,byref):System.Collections.IDictionary:this
         -15 (-2.74% of base) : 12882.dasm - Newtonsoft.Json.Serialization.JsonSerializerInternalReader:CreateNewList(Newtonsoft.Json.JsonReader,Newtonsoft.Json.Serialization.JsonArrayContract,byref):System.Collections.IList:this
         -13 (-2.11% of base) : 9473.dasm - System.Net.Http.HttpWindowsProxy:TryCreate(byref):bool
         -14 (-1.82% of base) : 18084.dasm - Microsoft.CodeAnalysis.CSharp.Symbols.Metadata.PE.PEParameterSymbol:.ctor(Microsoft.CodeAnalysis.CSharp.Symbols.Metadata.PE.PEModuleSymbol,Microsoft.CodeAnalysis.CSharp.Symbol,int,bool,Microsoft.CodeAnalysis.CSharp.Symbols.TypeSymbol,System.Reflection.Metadata.ParameterHandle,int,byref):this
         -18 (-1.36% of base) : 6098.dasm - JetStream.Statistics:findOptimalSegmentationInternal(System.Single[][],System.Int32[][],System.Double[],JetStream.SampleVarianceUpperTriangularMatrix,int)
         -13 (-1.06% of base) : 12880.dasm - Newtonsoft.Json.Serialization.JsonSerializerInternalReader:CreateList(Newtonsoft.Json.JsonReader,System.Type,Newtonsoft.Json.Serialization.JsonContract,Newtonsoft.Json.Serialization.JsonProperty,System.Object,System.String):System.Object:this
          -5 (-0.94% of base) : 13383.dasm - System.Net.Security.SslStream:ValidateCreateContext(System.Net.Security.SslClientAuthenticationOptions,System.Net.Security.RemoteCertificateValidationCallback,System.Net.Security.LocalCertSelectionCallback):this
         -13 (-0.89% of base) : 12412.dasm - Newtonsoft.Json.Serialization.JsonDictionaryContract:.ctor(System.Type):this
         -16 (-0.87% of base) : 18739.dasm - Microsoft.CodeAnalysis.CSharp.DataFlowPass:Scan(byref):System.Collections.Immutable.ImmutableArray`1[[Microsoft.CodeAnalysis.CSharp.PreciseAbstractFlowPass`1+PendingBranch[[Microsoft.CodeAnalysis.CSharp.DataFlowPass+LocalState, Microsoft.CodeAnalysis.CSharp, Version=2.10.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35]], Microsoft.CodeAnalysis.CSharp, Version=2.10.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35]]:this
          -2 (-0.80% of base) : 4913.dasm - System.Net.Sockets.SocketAsyncEventArgs:StartOperationAccept():this
          -2 (-0.50% of base) : 13451.dasm - System.Net.Security.SslStream:ValidateCreateContext(System.Net.Security.SslAuthenticationOptions):this
          -2 (-0.21% of base) : 18655.dasm - Microsoft.CodeAnalysis.CSharp.ConversionsBase:ClassifyImplicitBuiltInConversionFromExpression(Microsoft.CodeAnalysis.CSharp.BoundExpression,Microsoft.CodeAnalysis.CSharp.Symbols.TypeSymbol,Microsoft.CodeAnalysis.CSharp.Symbols.TypeSymbol,byref):Microsoft.CodeAnalysis.CSharp.Conversion:this
          -4 (-0.13% of base) : 1617.dasm - ProtoBuf.Meta.MetaType:BuildSerializer():ProtoBuf.Internal.Serializers.IProtoTypeSerializer:this
          -3 (-0.08% of base) : 18343.dasm - Microsoft.CodeAnalysis.CSharp.Binder:BindNamespaceOrTypeOrAliasSymbol(Microsoft.CodeAnalysis.CSharp.Syntax.ExpressionSyntax,Microsoft.CodeAnalysis.DiagnosticBag,Roslyn.Utilities.ConsList`1[[Microsoft.CodeAnalysis.CSharp.Symbol, Microsoft.CodeAnalysis.CSharp, Version=2.10.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35]],bool):Microsoft.CodeAnalysis.CSharp.Symbol:this

18 total methods with Code Size differences (17 improved, 1 regressed), 0 unchanged.

```

</details>

--------------------------------------------------------------------------------

## coreclr_tests.pmi.windows.x64.checked.mch:

```

Summary of Code Size diffs:
(Lower is better)

Total bytes of base: 124163274 (overridden on cmd)
Total bytes of diff: 124162492 (overridden on cmd)
Total bytes of delta: -782 (-0.00 % of base)
    diff is an improvement.
    relative diff is an improvement.
```
<details>

<summary>Detail diffs</summary>

```


Top file regressions (bytes):
           3 : 88633.dasm (4.05% of base)

Top file improvements (bytes):
        -222 : 249541.dasm (-14.61% of base)
         -77 : 83853.dasm (-32.08% of base)
         -51 : 88738.dasm (-61.45% of base)
         -49 : 250030.dasm (-4.24% of base)
         -30 : 84713.dasm (-14.42% of base)
         -30 : 248103.dasm (-10.60% of base)
         -30 : 85910.dasm (-10.27% of base)
         -30 : 248135.dasm (-10.27% of base)
         -30 : 237875.dasm (-10.27% of base)
         -30 : 237982.dasm (-10.60% of base)
         -15 : 88223.dasm (-5.79% of base)
         -15 : 88219.dasm (-7.50% of base)
         -15 : 88214.dasm (-6.98% of base)
         -13 : 209718.dasm (-1.98% of base)
         -13 : 209750.dasm (-0.23% of base)
         -12 : 209740.dasm (-0.70% of base)
         -12 : 209745.dasm (-2.82% of base)
         -11 : 84187.dasm (-2.72% of base)
          -8 : 88221.dasm (-3.65% of base)
          -4 : 102780.dasm (-0.61% of base)

44 total files with Code Size differences (43 improved, 1 regressed), 4 unchanged.

Top method regressions (bytes):
           3 ( 4.05% of base) : 88633.dasm - foo:Main():int

Top method improvements (bytes):
        -222 (-14.61% of base) : 249541.dasm - HelloWorld:Main(System.String[]):int
…